+++ plugins/pipeline_plugin/transform/default_transform.py
@@ -12,9 +12,7 @@ def transform(
     input_filename = load_file(input_filename)
 
     if source == "osm":
-        df_rail = transform_osm(
-            input_filename=input_filename, schema_mapping=schema_mapping
-        )
-        df_rail = postprocess(df=df_rail, crs=crs)
+        df = transform_osm(input_filename=input_filename, schema_mapping=schema_mapping)
+        df = postprocess(df=df, crs=crs)
 
-    save_shapefiles(df_rail, output_filename, encoding="utf8")
+    save_shapefiles(df, output_filename, encoding="utf8")
